refactor: migrate to layered directory structure with comprehensive tests
Major restructuring following NETWORK_DESIGN.md architecture:

## Directory Structure
- Migrate all source files to src/ hierarchy:
  - src/core/: chord_types.h, hash.*, key.*, ring.*, finger.*, node.*
  - src/util/: util.*
  - src/app/: app_driver.c
- Update all #include paths for new structure
- Fix .gitignore to allow src/core/ (was matching 'core' dumps)

## Code Quality Improvements
- Fix ALL compiler warnings to re-enable -Werror:
  - ring.c: Rename global 'ring' to 'g_ring' (shadowing fix)
  - ring.c: Fix sign conversion warnings (size_t casts)
  - ring.c: Remove unused variables
  - finger.c: Add explicit casts for pow() and malloc
  - node.c: Fix realloc size_t cast
  - key.c, node.c: Mark unused parameters with (void)
  - util.c: Fix fgets length cast
  - app_driver.c: Fix function signature, unused params, malloc cast
- All files now compile with -Werror -Wall -Wextra -Wpedantic

## Test Suite Expansion
- tests/unit/test_key.c: 6 tests for key_in_range()
  - Half-open and open interval tests
  - Wraparound behavior in circular keyspace
  - Edge cases and full-circle scenarios
- tests/unit/test_ring.c: 3 tests for ring operations
  - Singleton initialization
  - Keyspace size validation
  - Empty ring size check
- Update test_hash.c for new include paths
- All 15 unit tests passing (6 hash + 6 key + 3 ring)

## Build System
- Update Makefile for new src/ structure
- Add test-key and test-ring targets
- Use proper include paths (-Isrc/core -Isrc/util -Isrc/app)
- Maintain strict compiler flags with -Werror re-enabled

Test results: 15/15 passing ✓

+++ src/core/finger.c
@@ -27,12 +27,12 @@ FingerTable* finger_table_init(Node *node) {
   finger_table->length = KEY_BITS;
   
   /* allocate fingers in the table */
-  if ((finger_table->fingers = malloc(sizeof(Finger) * finger_table->length)) == NULL) {
+  if ((finger_table->fingers = malloc(sizeof(Finger) * (size_t)finger_table->length)) == NULL) {
     BAIL("Failed to allocate memory for Finger");
   }
   
   for (i = 0; i < finger_table->length; i++) {
-    start = node->key + pow(2, i);
+    start = node->key + (int)pow(2, i);
     if (start > ring_key_max()) {
       start -= ring_key_max();
     }
